project/requirements.txt
asgiref==3.7.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.6
cryptography==41.0.2
defusedxml==0.7.1
dj-paystack==0.1
Django==3.2.20
django-cors-headers==4.0.0
django-sslserver==0.22
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==21.2.0
httpie==3.2.2
idna==3.4
markdown-it-py==2.2.0
mdurl==0.1.2
multidict==6.0.4
packaging==23.1
Pillow==9.5.0
psycopg2-binary==2.9.7
pycparser==2.21
Pygments==2.15.1
PyJWT==2.7.0
PySocks==1.7.1
pytz==2023.3
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.4.2
sqlparse==0.4.4
stripe==4.2.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
whitenoise==6.2.0
Brotli==1.0.9
amqp==5.1.1
billiard==4.1.0
celery==5.3.4
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
kombu==5.3.2
vine==5.0.0
wcwidth==0.2.8
